fix: move checks to separate page
